split CLI commands into separate files
