+++ bin/teachers_pet
@@ -1,4 +1,4 @@
 #!/usr/bin/env ruby
 
-require_relative File.join('..', 'lib', 'teachers_pet', 'cli')
+require_relative File.join('..', 'lib', 'teachers_pet')
 TeachersPet::Cli.start(ARGV)
