+++ lib/teachers_pet.rb
@@ -1,3 +1,5 @@
+require 'active_support/concern'
+
 require 'require_all'
 require_rel '.'
